backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
playwright>=1.51.0,<2.0.0
sqlalchemy>=2.0.36,<3.0.0
aiosqlite>=0.20.0,<1.0.0
asyncpg>=0.30.0,<1.0.0
pydantic-settings>=2.6.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
celery>=5.4.0,<6.0.0
redis>=5.2.0,<6.0.0
pgvector>=0.3.6,<1.0.0
cloudinary>=1.40.0,<2.0.0
python-multipart>=0.0.18,<1.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
